split and review e2e tests
